requirements.txt
fastapi>=0.109.2
orjson>=3.9.14
nicegui>=1.4.15
httpx>=0.26.0
pywebview>=4.4.1
starlette>=0.36.3
Pillow>=10.2.0
markdown2>=2.4.12
Pygments>=2.17.2
uvicorn>=0.27.1
aiofiles>=23.2.1
ifaddr>=0.2.0
vbuild>=0.8.2
PyYAML>=6.0.1
setuptools>=69.1.0
requests>=2.31.0